Move the EstimateStride method to the BitmapData class.
