+++ FFMediaToolkit/src/Graphics/Layout.cs
@@ -54,11 +54,6 @@
         public Size Size { get; }
 
         /// <summary>
-        /// Gets the estimated image linesize.
-        /// </summary>
-        public int Stride => EstimateStride(Width, (ImagePixelFormat)PixelFormat);
-
-        /// <summary>
         /// Checks if the <paramref name="left"/> value equals the <paramref name="right"/> value.
         /// </summary>
         /// <param name="left">The left value.</param>
@@ -73,14 +68,6 @@
         /// <param name="right">The right value.</param>
         /// <returns><see langword="true"/>, if not equals.</returns>
         public static bool operator !=(Layout left, Layout right) => !(left == right);
-
-        /// <summary>
-        /// Gets the estimated image line size based on the pixel format and width.
-        /// </summary>
-        /// <param name="width">The image width.</param>
-        /// <param name="format">The image pixel format.</param>
-        /// <returns>The size of a single line of the image measured in bytes.</returns>
-        public static int EstimateStride(int width, ImagePixelFormat format) => GetBytesPerPixel(format) * width;
 
         /// <summary>
         /// Indicates whether the width and height of this layout are equal to the other.
@@ -103,23 +90,6 @@
             hashCode = (hashCode * -1521134295) + Width.GetHashCode();
             hashCode = (hashCode * -1521134295) + Height.GetHashCode();
             return hashCode;
-        }
-
-        private static int GetBytesPerPixel(ImagePixelFormat format)
-        {
-            switch (format)
-            {
-                case ImagePixelFormat.BGR24:
-                    return 3;
-                case ImagePixelFormat.BGRA32:
-                    return 4;
-                case ImagePixelFormat.RGB24:
-                    return 3;
-                case ImagePixelFormat.ARGB32:
-                    return 4;
-                default:
-                    return 0;
-            }
         }
     }
 }
